backend/requirements.txt
transformers
sentence-transformers
faiss-cpu
flask
flask-cors
torch
gunicorn
